Fundamental_of_Generative AI_for_Item_Development/Lectures/1.1 LangchainBasicTraining/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.8
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
jupyter>=1.0.0
streamlit>=1.28.0
chromadb>=0.4.0
pypdf>=3.0.0
requests>=2.31.0